+++ src/openapi-tsk/src/index.ts
@@ -1,4 +1,4 @@
-export { ApiDoc, ApiDocRouteType, IApiDocGenerator, SecurityScheme, SecuritySchemeType, UrlParamDescriber, ParameterIn } from "./resources/IApiDocGenerator";
+export { ApiDoc, ApiDocRouteType, IApiDocGenerator, SecurityScheme, SecuritySchemeType, UrlParamDescriber, ParameterIn, SchemeDescription, ApiProduce } from "./resources/IApiDocGenerator";
 export { TypeDescriber, RefTypeDescriber, SecuritySchemesDescriber } from "./resources/TypeDescriber";
 export { PropTypeEnum, PropFormatEnum, ClassProperty } from "./resources/types";
 export { SecuritySchemesStore } from "./resources/SecuritySchemesStore";