+++ src/openapi-tsk/CHANGELOG.md
@@ -3,3 +3,15 @@
 ## [1.0.0] — 2024-10-02
 
 - Released first version
+
+## [1.0.1] — 2024-10-04
+
+- Exported new types
+
+## [1.0.3] — 2024-10-04
+
+- Fixed apiKey security scheme
+
+## [1.0.4] — 2024-10-05
+
+- Option to support scheme and ref scheme in produces was added